+++ src/main/java/org/example/delivery/common/domain/User.java
@@ -1,0 +1,5 @@
+package org.example.delivery.common.domain;
+
+public class User {
+
+}
