--- src/main/java/org/example/delivery/common/entity/Menu.java
@@ -1,5 +1,0 @@
-package org.example.delivery.common.entity;
-
-public class Menu {
-
-}
