--- src/main/java/org/example/delivery/common/entity/Review.java
@@ -1,5 +1,0 @@
-package org.example.delivery.common.entity;
-
-public class Review {
-
-}
